tests/test_day04.py
import pathlib
import pytest
import sys

ROOT_DIR = pathlib.Path(__file__).parent.parent
sys.path.append(f"{ROOT_DIR}/puzzles")
import day04 as aoc

INPUTS_DIR = f"{ROOT_DIR}/inputs"

@pytest.fixture
def example_data():
    input_path = f"{INPUTS_DIR}/day04-example.txt"
    return aoc.parse(pathlib.Path(input_path).read_text().strip())


@pytest.fixture
def day04_data():
    input_path = f"{INPUTS_DIR}/day04.txt"
    return aoc.parse(pathlib.Path(input_path).read_text().strip())


def test_example1(example_data):
    assert aoc.part1(example_data) == 2


def test_example2(example_data):
    assert aoc.part2(example_data) == 4


def test_part1(day04_data):
    assert aoc.part1(day04_data) == 431


def test_part2(day04_data):
    assert aoc.part2(day04_data) == 823
